refactor: split remaining monoliths (preload, App, Plugins, Sync)
preload/index.ts: 1227 → 161 lines
- Extract 14 API modules into preload/api/ (notes, notebooks, sync, etc.)
- Each module exports typed factory function + interface
- Barrel re-exports all types for renderer compatibility

App.tsx: 1010 → 643 lines
- Extract useDeepLinks, useAutoSave, useNoteActions, useAppCommands hooks
- App.tsx becomes a shell: providers + layout + hook composition

PluginsSection.tsx: 1043 → 11 lines (barrel)
- Extract PluginCard, BrowseTab, PluginInspector components
- New plugins/ directory with types.ts and composed index.tsx

syncService.ts: 1065 → 18 lines (barrel)
- Extract types, helpers to sync/ directory
- Core SyncService class in its own file
- Auto-sync timer and conflict resolver kept inline (too coupled)

No behavior changes — purely structural.
