+++ apps/desktop/src/main/services/sync/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Sync module — re-exports for convenient importing.
+ */
+
+export { SyncService } from './SyncService.js';
+export type {
+  SyncConflict,
+  SyncResult,
+  SyncState,
+  SyncStatusEvent,
+  SyncStatusListener,
+} from './types.js';
